+++ src/main/java/com/example/QuizSystemProject/Repository/QuestionTypeRepository.java
@@ -10,7 +10,7 @@ import java.util.Optional; // Optional importu (Bulunamama durumuna karşı)
 @Repository // Spring'e bu arayüzün bir Repository olduğunu belirtir
 // JpaRepository'den türeyerek temel CRUD metotlarını alır
 // <QuestionType, Long>: İlk parametre Entity tipi (QuestionType), ikinci parametre Entity'nin Primary Key (ID) tipidir (Long).
-public interface QuestionTypeRepository extends JpaRepository<QuestionType, Long> {
+public interface QuestionTypeRepository extends JpaRepository<QuestionType, Integer> {
 
     // Spring Data JPA, JpaRepository'den türediğimiz için
     // save(), findById(), findAll(), delete() gibi temel metotları otomatik sağlar.
